Solved bug when discontinuous samples were to be stored.
If the number of samples to store was less than the segment size (due to
the occurrence of an event) the last buffer (which already belongs to
the new segment) was being incorrectly added to the old segment.

New test added to cover for this use-case. Code linted and coverage
checked.